update project web2

+++ app/Http/Controllers/CategoryController.php
@@ -5,6 +5,10 @@ namespace App\Http\Controllers;
 use Illuminate\Http\Request;
 use App\Models\categories;
 use App\Models\products;
+
+use function PHPUnit\Framework\isEmpty;
+
+use Illuminate\Support\Facades\Validator;
 class CategoryController extends Controller
 {
     /**
@@ -14,9 +18,13 @@ class CategoryController extends Controller
      */
     public function index()
     {
-        $products = products::all();
-        $categories = categories::all();
-        return $categories;
+        $categories = categories::all()->toArray();
+        $return = [];
+        foreach($categories as $item){
+            $item['id'] = $this->Xulyid($item['id']);
+            $return[] = $item;
+        }
+        return response()->json($return);
     }
 
     /**
@@ -26,7 +34,6 @@ class CategoryController extends Controller
      */
     public function create()
     {
-       
     }
 
     /**
@@ -37,6 +44,13 @@ class CategoryController extends Controller
      */
     public function store(Request $request)
     {
+        $validator = Validator::make($request->all(),[
+            'name'=>'required|unique:categories,name',
+        ]);
+
+            if ($validator->fails()){
+                return response(['errors'=>$validator->errors()->all()], 422);
+            }
         $category = new categories([
             'name' => $request->get('name'),
             'description' => $request->get('description'),
@@ -55,9 +69,20 @@ class CategoryController extends Controller
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function show(categories $category)
+    public function show(Request  $request,$id)
     {
-        return $category;
+
+        $cat_id = $this->DichId($id);
+        $category = categories::find($cat_id);
+        if ($category) {
+            return response()->json([
+                'message' => 'category found!',
+                'category' => $category,
+            ]);
+        }
+        return response()->json([
+            'message' => 'category not found!',
+        ]);
     }
 
 
@@ -68,12 +93,8 @@ class CategoryController extends Controller
      * @return \Illuminate\Http\Response
      */
     public function edit($id)
-    {    
-        $item = categories::find($id);
-        return response()->json([
-            'message' => 'Categories find it !!!',
-            'item' => $item
-        ]);
+    {
+
     }
 
     /**
@@ -85,25 +106,56 @@ class CategoryController extends Controller
      */
     public function update(Request $request, $id)
     {
-        /*$request->validate([
-            'category_name' => 'required',
-            'category_image' => ''
-        ]);*/
+        $list = categories::all()->toArray();
+        $return = [];
+        foreach($list as $item){
+            $item['id'] = $this->Xulyid($item['id']);
+            $return[] = $item;
+        }
 
-        //2 Tao Product Model, gan gia tri tu form len cac thuoc tinh cua category model
-        $category = categories::find($id);
-        $category->name = $request->get('name');
-        $category->description = $request->get('description');
-        $category->image = "";
+      $cat_id = $this->DichId($id);
+      $category = categories::find($cat_id);
 
 
-        //3 Luu
+      if($category){
+      $validator = Validator::make($request->all(),[
+        'name'=>'required',
+    ]);
+
+        if ($validator->fails()){
+            return response(['errors'=>$validator->errors()->all()], 422);
+        }
+
+        $cat_name;
+
+        //Kiem tra name da co hay chua, co bi trung khong
+        if($request->name == $list[0]['name']){
+            return response()->json([
+                'message' => 'The name has been exits!!!',
+            ]);
+        }else{
+            $cat_name = $request->name;
+        }
+
+        $category->update([
+        $category->name = $cat_name,
+        // $category->image = $request->get('image');
+        ]);
+
+
         $category->save();
         return response()->json([
-            'message' => 'categories updated successfully !!!',
-            'category' => $category,
+            'message' => 'category updated!',
+            'category' => $category
+        ]);
+
+      }
+
+        return response()->json([
+            'message' => 'category not found !!!'
         ]);
     }
+
 
     /**
      * Remove the specified resource from storage.
@@ -113,18 +165,68 @@ class CategoryController extends Controller
      */
     public function destroy($id)
     {
+        $id = $this->DichId($id);
+        $pattern_id = '/^\d{1,}$/';
+        if (!preg_match($pattern_id,$id)){
+            return  response()->json(['message'=>'Please type id is a number']);
+        }
         $category = categories::find($id);
-        $category->delete();
+        $productListTemp =  products::where("category_id", "=", $id)->get();
+        if (count($productListTemp) === 0) {
+            $category->delete();
+            return response()->json([
+                'message' => 'categories deleted successfully !!!',
+                'item' => $category
+            ]);
+        }
         return response()->json([
-            'message' => 'categories deleted successfully !!!',
-            'item' => $category
+            'message' => "can't delete because have a dependent products",
         ]);
     }
+
+    private function getName($n) {
+        $characters = '162379812362378dhajsduqwyeuiasuiqwy460123';
+        $randomString = '';
+
+        for ($i = 0; $i < $n; $i++) {
+            $index = rand(0, strlen($characters) - 1);
+            $randomString .= $characters[$index];
+        }
+        return $randomString;
+    }
+
+    public  function Xulyid($id):String {
+        $dodaichuoi = strlen($id);
+        $chuoitruoc = $this->getName(10);
+        $chuoisau = $this->getName(22);
+        $handle_id = base64_encode($chuoitruoc.$id. $chuoisau);
+        return $handle_id;
+    }
+
+    public function DichId($id){
+        $id = base64_decode($id);
+        $handleFirst = substr($id,10);
+        $idx = "";
+        for ($i=0; $i <strlen($handleFirst)-22 ; $i++) {
+            $idx.=$handleFirst[$i];
+        }
+        return  $idx;
+    }
+
+
     public function getSearch(Request $request){
-        $category = categories::where('name','like','%'.$request->keyword.'%')->get();
-        return response()->json([
-            'message' => 'categories find it !!!',
-            'item' => $category
-        ]);
+        $category = categories::where('name','like','%'.$request->key.'%')->get();
+        if($category){
+            if(empty(count($category))){
+                return response()->json([
+                    'message' => 'category not found!',
+                ]);
+            } else {
+                return response()->json([
+                    'message' => count($category) . ' category found!!!',
+                    'item' => $category
+                ]);
+            }
+        }
     }
 }
